+++ src/entities/comment/model/types.ts
@@ -1,0 +1,10 @@
+import { User } from "../../user/model/types"
+
+export interface Comment {
+  id: number
+  postId: number
+  userId: number
+  body: string
+  likes: number
+  user?: User
+}
