+++ src/entities/post/model/types.ts
@@ -1,0 +1,21 @@
+import { User } from "../../user/model/types"
+
+export interface Post {
+  id: number
+  title: string
+  body: string
+  userId: number
+  tags?: string[]
+  author?: User
+  reactions?: Reactions
+}
+
+interface Reactions {
+  likes: number
+  dislikes: number
+}
+
+export interface FetchPostsResponse {
+  posts: Post[]
+  total: number
+}
