+++ src/shared/model/types.ts
@@ -1,0 +1,4 @@
+export interface Tag {
+  url: string
+  slug: string
+}
